backend/requirements.txt
fastapi
uvicorn
pydantic
openai
google-generativeai
anthropic
python-docx
pandas
python-multipart
beautifulsoup4
requests
lxml
sqlalchemy>=2.0
psycopg2-binary
alembic
python-dotenv
bcrypt
pyjwt
python-multipart
